move animation classes into their own package
