+++ core/src/com/mygdx/game/animations/CharacterAnimationType.java
@@ -1,4 +1,4 @@
-package com.mygdx.game;
+package com.mygdx.game.animations;
 
 public enum CharacterAnimationType {
   MAGEIDLE,
