+++ core/src/com/mygdx/game/animations/EnemyAnimation.java
@@ -1,4 +1,4 @@
-package com.mygdx.game;
+package com.mygdx.game.animations;
 
 import com.badlogic.gdx.Gdx;
 import com.badlogic.gdx.graphics.Texture;
